requirements.txt
altgraph==0.10.2
bdist-mpkg==0.5.0
bonjour-py==0.3
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.7.3
django-filter==0.9.1
django-toolbelt==0.0.1
djangorestframework==3.0.3
gunicorn==19.1.1
macholib==1.5.1
Markdown==2.5.2
matplotlib==1.3.1
modulegraph==0.10.4
numpy==1.8.0rc1
psycopg2==2.5.4
py2app==0.7.3
pyOpenSSL==0.13.1
pyparsing==2.0.1
python-dateutil==1.5
pytz==2013.7
scipy==0.13.0b1
six==1.4.1
static3==0.5.1
Twisted==13.2.0
xattr==0.6.4
zope.interface==4.1.1
